refactor: move fertilizePlot to its own file
