+++ src/game-logic/reducers/field.js
@@ -38,65 +38,12 @@ import { modifyFieldPlotAt } from './modifyFieldPlotAt'
 import { removeFieldPlotAt } from './removeFieldPlotAt'
 import { plantInPlot } from './plantInPlot'
 
-const { FERTILIZE, OBSERVE, SET_SCARECROW, SET_SPRINKLER } = fieldMode
+const { OBSERVE, SET_SCARECROW, SET_SPRINKLER } = fieldMode
 const { GROWN } = cropLifeStage
 
-const fertilizerItemIdToTypeMap = {
+export const fertilizerItemIdToTypeMap = {
   [itemsMap['fertilizer'].id]: fertilizerType.STANDARD,
   [itemsMap['rainbow-fertilizer'].id]: fertilizerType.RAINBOW,
-}
-
-/**
- * Assumes that state.selectedItemId references an item with type ===
- * itemType.FERTILIZER.
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @returns {farmhand.state}
- */
-export const fertilizePlot = (state, x, y) => {
-  const { field, selectedItemId } = state
-  const row = field[y]
-  const plotContent = row[x]
-
-  if (!plotContent || itemsMap[selectedItemId]?.type !== itemType.FERTILIZER) {
-    return state
-  }
-
-  const { id: fertilizerItemId } = itemsMap[selectedItemId]
-
-  const fertilizerInventory = state.inventory.find(
-    item => item.id === fertilizerItemId
-  )
-
-  const plotContentType = getPlotContentType(plotContent)
-
-  if (
-    !plotContent ||
-    !fertilizerInventory ||
-    plotContent.fertilizerType !== fertilizerType.NONE ||
-    (selectedItemId === 'fertilizer' && plotContentType !== itemType.CROP) ||
-    (selectedItemId === 'rainbow-fertilizer' &&
-      plotContentType !== itemType.CROP &&
-      plotContentType !== itemType.SCARECROW)
-  ) {
-    return state
-  }
-
-  const { quantity: initialFertilizerQuantity } = fertilizerInventory
-  state = decrementItemFromInventory(state, fertilizerItemId)
-  const doFertilizersRemain = initialFertilizerQuantity > 1
-
-  state = modifyFieldPlotAt(state, x, y, crop => ({
-    ...crop,
-    fertilizerType: fertilizerItemIdToTypeMap[fertilizerItemId],
-  }))
-
-  return {
-    ...state,
-    fieldMode: doFertilizersRemain ? FERTILIZE : OBSERVE,
-    selectedItemId: doFertilizersRemain ? fertilizerItemId : '',
-  }
 }
 
 /**
