+++ src/game-logic/reducers/index.js
@@ -118,6 +118,7 @@ export * from './processSprinklers'
 export * from './modifyFieldPlotAt'
 export * from './removeFieldPlotAt'
 export * from './plantInPlot'
+export * from './fertilizePlot'
 
 ///////////////////////////////////////////////////////////
 //
